[feat] FT-37 Router를 Public과 Private로 분류하기
