+++ srcs/front/src/modules/login/isAdmin.ts
@@ -1,0 +1,5 @@
+const isAdmin = () => {
+	return !!localStorage.getItem('token');
+};
+
+export default isAdmin;
