+++ srcs/front/src/routes/Router.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import HomePage from './HomePage';
-import LoginPage from './LoginPage';
-import AuthPage from './AuthPage';
-import GamePage from './GamePage';
-import ProfilePage from './ProfilePage';
-import FriendsPage from './FriendsPage';
-import LeaderboardPage from './LeaderboardPage';
+import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
+import HomePage from './pages/HomePage';
+import LoginPage from './pages/LoginPage';
+import AuthPage from './pages/AuthPage';
+import GamePage from './pages/GamePage';
+import ProfilePage from './pages/ProfilePage';
+import FriendsPage from './pages/FriendsPage';
+import LeaderboardPage from './pages/LeaderboardPage';
+import ErrorPage from './pages/ErrorPage';
+import isAdmin from '../modules/login/isAdmin';
 
 function Router() {
 	return (
@@ -15,10 +17,11 @@ function Router() {
 				<Route path="/" element={<HomePage />} />
 				<Route path="/login" element={<LoginPage />} />
 				<Route path="/auth" element={<AuthPage />} />
-				<Route path="/game" element={<GamePage />} />
-				<Route path="/profile" element={<ProfilePage />} />
-				<Route path="/friends" element={<FriendsPage />} />
-				<Route path="/leaderboard" element={<LeaderboardPage />} />
+				<Route path="/game" element={isAdmin() ? <GamePage /> : <Navigate to="/login" />} />
+				<Route path="/profile" element={isAdmin() ? <ProfilePage /> : <Navigate to="/login" />} />
+				<Route path="/friends" element={isAdmin() ? <FriendsPage /> : <Navigate to="/login" />} />
+				<Route path="/leaderboard" element={isAdmin() ? <LeaderboardPage /> : <Navigate to="/login" />} />
+				<Route path="*" element={<ErrorPage />} />
 			</Routes>
 		</BrowserRouter>
 	);
