+++ srcs/front/src/routes/pages/ErrorPage.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ErrorPage() {
+	return (
+		<div>
+			<h1>ErrorPage</h1>
+		</div>
+	);
+}
+
+export default ErrorPage;
